HD_Xray_Pretrain_MAE/finetune/DP/requirements.txt
scipy
torch
torchvision
tqdm
easydict
numpy
Pillow
sentence_transformers
torchmetrics
scikit-learn